requirements.txt
discord.py==2.2.2
Requests==2.29.0
